refactor: внес изменения в компонент resultField. Теперь компоненты resultField и inputField подключаюются отдельно.
