+++ src/components/resultField/resultField.tsx
@@ -1,13 +1,15 @@
 import {Card}from '@components/card/card';
 import {usePagination} from '@utils/hooks/usePagination';
 import resultFieldStyles from "./resultField.module.css";
+import { useSelector } from 'react-redux';
 
-type ResultFieldProps = {
-  movies: object[];
-}
+// type ResultFieldProps = {
+//   movies: object[];
+// }
 
-export const ResultField: React.FC<ResultFieldProps> = ({ movies }) => {
-  const { currentMovies, currentPage, nextPage, prevPage } = usePagination(movies);
+export const ResultField: React.FC= () => {
+  const movies = useSelector(state => state.movies.movies);
+  const { currentMovies, currentPage, nextPage, prevPage } = usePagination();
   return (
     <div className={resultFieldStyles.resultContainer}>
       <div className={resultFieldStyles.resultField}>
